ADFA-2435 | Host wizard as Fragment; wrap inflater in plugin fragments
Plugin APKs load via DexClassLoader, so any <activity> in this plugin's
manifest never enters the host's merged AndroidManifest. Intent(ctx,
WizardActivity::class.java) was guaranteed to throw ActivityNotFoundException
in an actually-installed plugin, making the wizard unreachable.

Refactor the wizard into a Fragment hosted inside the schema panel's editor
tab via childFragmentManager:

- WizardHostFragment owns the chrome (title + back/cancel/continue) and
  hosts Step1..Step4 via its own childFragmentManager. Step fragments
  scope their shared WizardViewModel to the host fragment.
- SchemaPanelFragment becomes a thin host. Default child is the new
  SchemaPanelContentFragment (the previous panel UI); the Capture button
  pushes WizardHostFragment onto the panel's back stack. Cancel/Finish
  pops back to the content fragment, which refreshes the schema on resume.
- The toolbar action no longer launches the wizard; it just opens the
  panel tab and the user taps Capture there. The wizard has no entry
  point of its own anymore.
- Drop the plugin-declared <activity>, the WizardActivity class, and the
  activity_forms_wizard layout. The fragment-equivalent layout
  fragment_forms_wizard.xml replaces it.

Also wrap the layout inflater via PluginFragmentHelper.getPluginInflater
in every fragment + DialogFragment so resource lookups resolve against
the plugin APK's resource table, not the host's. Required for
SchemaPanelFragment when surfaced as an editor tab; required for the
wizard step fragments and FieldEditorDialog now that they're no longer
inflated from inside a plugin-APK Activity scope.

Also add FormsPlugin.PLUGIN_ID companion constant matching the
plugin.id manifest meta-data, so fragments don't have to hard-code the
string.

Build is green and the existing 30 JVM unit tests continue to pass.
Smoke-testing against an installed host build is still recommended:
JVM tests don't exercise the plugin loader or the resource path.

Refs B1, B2 from REVIEW.md and C1, C2 from REVIEW2.md.
