+++ forms-plugin/src/main/kotlin/com/appdevforall/forms/plugin/wizard/Step2ReviewFieldsFragment.kt
@@ -5,12 +5,14 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import androidx.fragment.app.activityViewModels
+import androidx.fragment.app.viewModels
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.appdevforall.forms.plugin.FormField
+import com.appdevforall.forms.plugin.FormsPlugin
 import com.appdevforall.forms.plugin.R
 import com.google.android.material.button.MaterialButton
+import com.itsaky.androidide.plugins.base.PluginFragmentHelper
 
 /**
  * Step 2 of 4 — Review fields.
@@ -22,8 +24,17 @@ import com.google.android.material.button.MaterialButton
  */
 class Step2ReviewFieldsFragment : Fragment() {
 
-    private val viewModel: WizardViewModel by activityViewModels()
+    private val viewModel: WizardViewModel by viewModels(
+        ownerProducer = { requireParentFragment() },
+    )
     private lateinit var adapter: FieldsAdapter
+
+    override fun onGetLayoutInflater(savedInstanceState: Bundle?): LayoutInflater {
+        return PluginFragmentHelper.getPluginInflater(
+            FormsPlugin.PLUGIN_ID,
+            super.onGetLayoutInflater(savedInstanceState),
+        )
+    }
 
     override fun onCreateView(
         inflater: LayoutInflater,
