+++ forms-plugin/src/main/kotlin/com/appdevforall/forms/plugin/wizard/Step3RulesFragment.kt
@@ -5,10 +5,12 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import androidx.fragment.app.activityViewModels
+import androidx.fragment.app.viewModels
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
+import com.appdevforall.forms.plugin.FormsPlugin
 import com.appdevforall.forms.plugin.R
+import com.itsaky.androidide.plugins.base.PluginFragmentHelper
 
 /**
  * Step 3 of 4 — Semantic rules.
@@ -18,8 +20,17 @@ import com.appdevforall.forms.plugin.R
  */
 class Step3RulesFragment : Fragment() {
 
-    private val viewModel: WizardViewModel by activityViewModels()
+    private val viewModel: WizardViewModel by viewModels(
+        ownerProducer = { requireParentFragment() },
+    )
     private lateinit var adapter: RulesAdapter
+
+    override fun onGetLayoutInflater(savedInstanceState: Bundle?): LayoutInflater {
+        return PluginFragmentHelper.getPluginInflater(
+            FormsPlugin.PLUGIN_ID,
+            super.onGetLayoutInflater(savedInstanceState),
+        )
+    }
 
     override fun onCreateView(
         inflater: LayoutInflater,
