+++ forms-plugin/src/main/res/layout/fragment_forms_wizard.xml
