+++ forms-plugin/src/main/res/layout/fragment_schema_panel_content.xml
