+++ forms-plugin/src/main/res/layout/fragment_schema_panel_host.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:id="@+id/forms_panel_host_container"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent" />
